% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_features}
\alias{extract_features}
\title{Extract features}
\usage{
extract_features(mod)
}
\arguments{
\item{mod}{Coxph model object fitted with coxph (survival).}
}
\value{
Features
}
\description{
Extract relevant features from a glmnet cox fit object.
}
\references{
Noah Simon, Jerome Friedman, Trevor Hastie, Rob Tibshirani (2011).
  Regularization Paths for Cox's Proportional Hazards Model via
 Coordinate Descent. Journal of Statistical Software, 39(5), 1-13.
 URL  http://www.jstatsoft.org/v39/i05/.
}
\seealso{
[glmnet]
}
\author{
Carlos S Traynor
}
\keyword{glmnet}
