% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gen_stan_dat}
\alias{gen_stan_dat}
\title{Fit MAP estimates for the Poisson Model PEM.   Important aspects of the modelling #'approach: Dependent variable is the binary event indicator. Adding an intercept is #'recommendable to provide a decent and interpretable reference category for factor #'variables. Use the logarithm of the interval length as an offset.}
\usage{
gen_stan_dat(dat, status = "status", time = "time")
}
\arguments{
\item{d}{a dataset \cr
formula  \cr
a log baseline hazard
t_dur duration time for individtual ith in the hth interval
lambda hazard rate
os_event censoring indicator (actual event) for the ith subject in the hth interval}
}
\value{
a MAP fit
}
\description{
Fit MAP estimates for the Poisson Model PEM.   Important aspects of the modelling #'approach: Dependent variable is the binary event indicator. Adding an intercept is #'recommendable to provide a decent and interpretable reference category for factor #'variables. Use the logarithm of the interval length as an offset.
}
