% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_names}
\alias{gene_names}
\title{Gene Names}
\value{
\item{gene_names}{a data frame}
}
\description{
Gene Names
}
\details{
These data are the Hugo Symbol and Entrez Id for the Microarray data
and CNA experiment. It matches Hugo Symbol and Entrez Id.
}
\examples{
data(gene_names)
str(gene_names)
}
\references{
Pereira, B. et al. Nat. Commun. 7, 2016.
}
\keyword{datasets}
