% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iclust2_glmnet}
\alias{iclust2_glmnet}
\title{iClust2 GLMNET}
\value{
\item{iclust2_glmnet}{a glmnet object}
}
\description{
iClust2 GLMNET
}
\details{
These data are the cross-validation elastic net fit object for
only using intClust2 patients.
}
\examples{
data(iclust2_glmnet)
print(iclust2_glmnet)
plot(iclust2_glmnet)
}
\references{
Noah Simon, Jerome Friedman, Trevor Hastie, Rob Tibshirani (2011).
 Regularization Paths for Cox's Proportional Hazards Model via Coordinate
 Descent. Journal of Statistical Software, 39(5), 1-13.
 URL http://www.jstatsoft.org/v39/i05/.
}
\keyword{datasets}
