% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intclust2dat}
\alias{intclust2dat}
\title{intClust2 data}
\value{
\item{intclust2dat}{a data frame}
}
\description{
intClust2 data
}
\details{
These data are the response variables time/status and only includs
 selected variables by glmnet, either using intclust2 only or the pooled dataset.
}
\examples{
data(intclust2dat)
str(intclust2dat)
}
\references{
Pereira, B. et al. Nat. Commun. 7, 2016.
}
\keyword{datasets}
