% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{par.table}
\alias{par.table}
\title{Function par table}
\usage{
par.table(fit, log_scale = FALSE)
}
\arguments{
\item{obj}{: survival coxph fit. \cr}
}
\value{
a coxph fit object
}
\description{
Extract features of a classical fit model
}
\author{
Sahota Tarj - caret
}
