% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pooled_glmnet}
\alias{pooled_glmnet}
\title{pooled GLMNET}
\value{
\item{pooled_glmnet}{a glmnet object}
}
\description{
pooled GLMNET
}
\details{
These data are the cross-validation elastic net fit object for
the pooled dataset.
}
\examples{
data(pooled_glmnet)
print(pooled_glmnet)
plot(pooled_glmnet)
}
\references{
Noah Simon, Jerome Friedman, Trevor Hastie, Rob Tibshirani (2011).
 Regularization Paths for Cox's Proportional Hazards Model via Coordinate
 Descent. Journal of Statistical Software, 39(5), 1-13.
 URL http://www.jstatsoft.org/v39/i05/.
}
\keyword{datasets}
