% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{post_tab}
\alias{post_tab}
\title{Function posterior table}
\usage{
post_tab(diff_tab, ibrier_tab)
}
\arguments{
\item{obj}{: posterior distribution and contrasts \cr}
}
\value{
a table for publication
}
\description{
Extract features of a Bayesian model
}
